Added Error checking for Rollover date
If current real date is 10-Nov-2021, at 2259, and I want to book today. I will be able to finish "askUserForDate"
But after entering "askUserForTime", and if the real time rolls over to 2300, I will no longer be able to book. So error checking checks for that and returns invalid values, creating a  DateTimeException, which causes "askUserForTime" function to exit prematurely.